alim

+++ panier.php
@@ -67,10 +67,11 @@ echo "bonjour monsieur  : ".$_SESSION['nom']." ".$_SESSION['prenom'];
 
 }
 //supprission de recette
+if(isset($_GET['titre'])){
 $titreRecetteAsupp=$_GET['titre'];
 $suppRecette="DELETE FROM recette WHERE   titre like '$titreRecetteAsupp' ";
 $database->query($suppRecette)->fetch();
-
+}
 
 //nombre de recette d'un PANIER
